// ordergroove/types/paths/subscriptions/{subscription_id}/change_quantity.types.ts
// This code was automatically generated from an OpenAPI description.
// Do not edit this file. Edit the OpenAPI file instead.
// For more information, see https://github.com/pmcelhaney/counterfact/blob/main/docs/faq.md

import type { WideOperationArgument } from "../../../../counterfact-types/index.ts";
import type { OmitValueWhenNever } from "../../../../counterfact-types/index.ts";
import type { MaybePromise } from "../../../../counterfact-types/index.ts";
import type { COUNTERFACT_RESPONSE } from "../../../../counterfact-types/index.ts";
import type { Context } from "../../../../routes/_.context.ts";
import type { ResponseBuilderFactory } from "../../../../counterfact-types/index.ts";
import type { Subscription } from "../../../components/schemas/Subscription.js";
import type { BadRequest } from "../../../#/components/responses/BadRequest.js";
import type { AuthenticationFailed } from "../../../#/components/responses/AuthenticationFailed.js";
import type { NotFound } from "../../../#/components/responses/NotFound.js";

/**
 * Updates quantity; prepaid subscriptions return 400. Source reviewed 2026-08-02:
 * https://developer.ordergroove.com/reference/subscriptions-change-quantity
 *
 */
export type subscriptionsChangeQuantity = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: subscriptionsChangeQuantity_Path;
    headers: subscriptionsChangeQuantity_Headers;
    cookie: never;
    body: { quantity: number };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Subscription;
          };
        };
        examples: {};
      };
      400: BadRequest;
      403: AuthenticationFailed;
      404: NotFound;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type subscriptionsChangeQuantity_Path = { subscription_id: string };

export type subscriptionsChangeQuantity_Headers = {
  "x-api-key": string;
  Authorization: string;
};